Dark mode submenu + AA monochrome filter
Move all dark mode settings into a dedicated sub-screen to
reduce clutter on the main settings page.  Add a new
"Make Android Auto Monochrome" toggle that applies a real-time
grayscale shader (GLES) to the AA video feed when night mode
is active, with an adjustable desaturation slider.
