+++ app/src/main/res/values-cs/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">Aplikace kreslí vlastní pozadí s přechodovým kruhem uprostřed. Neplatí pro režim Extrémně tmavý</string>
     <string name="use_gradient_background_description_off_auto">Používá se klasické pozadí aplikace. Neplatí pro režim Extrémně tmavý</string>
     <string name="category_dark_mode">Tmavý režim</string>
+    <string name="dark_mode_settings">Nastavení tmavého režimu</string>
+    <string name="aa_monochrome">Černobílý Android Auto</string>
+    <string name="aa_monochrome_description">Snižuje barvy Android Auto pro noční jízdu. Vyžaduje režim zobrazení GLES. Projeví se pouze při zapnutém tmavém režimu</string>
+    <string name="aa_desaturation">Desaturace</string>
+    <string name="aa_desaturation_description">Určuje, kolik barvy se odstraní. Projeví se pouze při zapnutém tmavém režimu</string>
+    <string name="gles_required_title">Vyžadován GLES</string>
+    <string name="gles_required_message">Monochromatický režim potřebuje vykreslování GLES pro použití filtru ve stupních šedi na video výstup. Povolit GLES?</string>
+    <string name="enable_gles">Povolit GLES</string>
     <string name="category_navigation">Navigace</string>
 
     <string name="reconnection_required_title">Vyžadováno opětovné připojení</string>
