+++ app/src/main/res/values-de/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">Die App zeichnet ihren eigenen Hintergrund mit dem zentrierten Verlaufskreis. Gilt nicht für den Modus Extrem Dunkel</string>
     <string name="use_gradient_background_description_off_auto">Der klassische Hintergrund der App wird verwendet. Gilt nicht für den Modus Extrem Dunkel</string>
     <string name="category_dark_mode">Dunkelmodus</string>
+    <string name="dark_mode_settings">Dunkelmodus-Einstellungen</string>
+    <string name="aa_monochrome">Android Auto einfarbig darstellen</string>
+    <string name="aa_monochrome_description">Reduziert die Farben von Android Auto für Nachtfahrten. Erfordert den GLES-Anzeigemodus. Wirkt nur bei aktiviertem Dunkelmodus</string>
+    <string name="aa_desaturation">Entsättigung</string>
+    <string name="aa_desaturation_description">Steuert, wie viel Farbe entfernt wird. Wirkt nur bei aktiviertem Dunkelmodus</string>
+    <string name="gles_required_title">GLES erforderlich</string>
+    <string name="gles_required_message">Der Monochrom-Modus benötigt GLES-Rendering, um einen Graustufen-Filter auf die Videoausgabe anzuwenden. GLES aktivieren?</string>
+    <string name="enable_gles">GLES aktivieren</string>
     <string name="category_navigation">Navigation</string>
 
     <string name="reconnection_required_title">Neuverbindung erforderlich</string>
