+++ app/src/main/res/values-es/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">La app dibuja su propio fondo con el círculo de degradado centrado. No aplica al modo Oscuro extremo</string>
     <string name="use_gradient_background_description_off_auto">Se está usando el fondo clásico de la app. No aplica al modo Oscuro extremo</string>
     <string name="category_dark_mode">Modo oscuro</string>
+    <string name="dark_mode_settings">Ajustes del modo oscuro</string>
+    <string name="aa_monochrome">Android Auto monocromático</string>
+    <string name="aa_monochrome_description">Reduce los colores de Android Auto para la conducción nocturna. Requiere el modo de vista GLES. Solo tiene efecto cuando el modo oscuro está activado</string>
+    <string name="aa_desaturation">Desaturación</string>
+    <string name="aa_desaturation_description">Controla cuánto color se elimina. Solo tiene efecto cuando el modo oscuro está activado</string>
+    <string name="gles_required_title">Se requiere GLES</string>
+    <string name="gles_required_message">El modo monocromático necesita renderizado GLES para aplicar un filtro de escala de grises a la salida de vídeo. ¿Activar GLES?</string>
+    <string name="enable_gles">Activar GLES</string>
     <string name="category_navigation">Navegación</string>
 
     <string name="reconnection_required_title">Reconexión necesaria</string>
