+++ app/src/main/res/values-hu/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">Az alkalmazás saját hátteret rajzol a középre igazított színátmenetes körrel. Nem vonatkozik az Extrém sötét módra</string>
     <string name="use_gradient_background_description_off_auto">Az alkalmazás klasszikus háttere van használatban. Nem vonatkozik az Extrém sötét módra</string>
     <string name="category_dark_mode">Sötét mód</string>
+    <string name="dark_mode_settings">Sötét mód beállítások</string>
+    <string name="aa_monochrome">Android Auto monokróm megjelenítés</string>
+    <string name="aa_monochrome_description">Csökkenti az Android Auto színeit az éjszakai vezetéshez. GLES megjelenítési mód szükséges. Csak bekapcsolt sötét módban lép érvénybe</string>
+    <string name="aa_desaturation">Színtelenítés</string>
+    <string name="aa_desaturation_description">Szabályozza, mennyi szín kerüljön eltávolításra. Csak bekapcsolt sötét módban lép érvénybe</string>
+    <string name="gles_required_title">GLES szükséges</string>
+    <string name="gles_required_message">A monokróm mód GLES renderelést igényel a szürkeárnyalatos szűrő alkalmazásához a videokimeneten. Engedélyezed a GLES-t?</string>
+    <string name="enable_gles">GLES engedélyezése</string>
     <string name="category_navigation">Navigáció</string>
 
     <string name="reconnection_required_title">Újrakapcsolódás szükséges</string>
