+++ app/src/main/res/values-nl/strings.xml
@@ -41,6 +41,14 @@
     <string name="use_gradient_background_description_on_auto">De app tekent een eigen achtergrond met de gecentreerde verloopcirkel. Geldt niet voor de modus Extreem donker</string>
     <string name="use_gradient_background_description_off_auto">De klassieke achtergrond van de app wordt gebruikt. Geldt niet voor de modus Extreem donker</string>
     <string name="category_dark_mode">Donkere modus</string>
+    <string name="dark_mode_settings">Instellingen donkere modus</string>
+    <string name="aa_monochrome">Android Auto monochroom maken</string>
+    <string name="aa_monochrome_description">Vermindert de kleuren van Android Auto voor nachtelijk rijden. Vereist GLES-weergavemodus. Werkt alleen wanneer donkere modus is ingeschakeld</string>
+    <string name="aa_desaturation">Desaturatie</string>
+    <string name="aa_desaturation_description">Bepaalt hoeveel kleur wordt verwijderd. Werkt alleen wanneer donkere modus is ingeschakeld</string>
+    <string name="gles_required_title">GLES vereist</string>
+    <string name="gles_required_message">Monochrome modus heeft GLES-rendering nodig om een grijstintenfilter toe te passen op de video-uitvoer. GLES inschakelen?</string>
+    <string name="enable_gles">GLES inschakelen</string>
     <string name="category_navigation">Navigatie</string>
 
     <string name="reconnection_required_title">Opnieuw verbinden vereist</string>
