+++ app/src/main/res/values-pt-rBR/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">O app desenha seu próprio fundo com o círculo de gradiente centralizado. Não se aplica ao modo Escuro extremo</string>
     <string name="use_gradient_background_description_off_auto">O fundo clássico do app está sendo usado. Não se aplica ao modo Escuro extremo</string>
     <string name="category_dark_mode">Modo escuro</string>
+    <string name="dark_mode_settings">Configurações do modo escuro</string>
+    <string name="aa_monochrome">Android Auto monocromático</string>
+    <string name="aa_monochrome_description">Reduz as cores do Android Auto para direção noturna. Requer o modo de visualização GLES. Só tem efeito quando o modo escuro está ativado</string>
+    <string name="aa_desaturation">Dessaturação</string>
+    <string name="aa_desaturation_description">Controla quanta cor é removida. Só tem efeito quando o modo escuro está ativado</string>
+    <string name="gles_required_title">GLES necessário</string>
+    <string name="gles_required_message">O modo monocromático precisa da renderização GLES para aplicar um filtro de escala de cinza na saída de vídeo. Ativar GLES?</string>
+    <string name="enable_gles">Ativar GLES</string>
     <string name="category_navigation">Navegação</string>
 
     <string name="reconnection_required_title">Reconexão necessária</string>
