+++ app/src/main/res/values-ro/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">Aplicația desenează propriul fundal cu cercul de degradat centrat. Nu se aplică modului Întunecat extrem</string>
     <string name="use_gradient_background_description_off_auto">Fundalul clasic al aplicației este utilizat. Nu se aplică modului Întunecat extrem</string>
     <string name="category_dark_mode">Mod întunecat</string>
+    <string name="dark_mode_settings">Setări mod întunecat</string>
+    <string name="aa_monochrome">Android Auto monocrom</string>
+    <string name="aa_monochrome_description">Reduce culorile din Android Auto pentru condusul pe timp de noapte. Necesită modul de vizualizare GLES. Are efect doar când modul întunecat este activat</string>
+    <string name="aa_desaturation">Desaturare</string>
+    <string name="aa_desaturation_description">Controlează cât de multă culoare este eliminată. Are efect doar când modul întunecat este activat</string>
+    <string name="gles_required_title">GLES necesar</string>
+    <string name="gles_required_message">Modul monocrom necesită randare GLES pentru a aplica un filtru de tonuri de gri pe ieșirea video. Activezi GLES?</string>
+    <string name="enable_gles">Activează GLES</string>
     <string name="category_navigation">Navigare</string>
 
     <string name="reconnection_required_title">Reconectare necesară</string>
