+++ app/src/main/res/values-ru/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">Приложение рисует собственный фон с центрированным градиентным кругом. Не применяется к режиму Экстремально тёмный</string>
     <string name="use_gradient_background_description_off_auto">Используется классический фон приложения. Не применяется к режиму Экстремально тёмный</string>
     <string name="category_dark_mode">Тёмный режим</string>
+    <string name="dark_mode_settings">Настройки тёмного режима</string>
+    <string name="aa_monochrome">Монохромный Android Auto</string>
+    <string name="aa_monochrome_description">Уменьшает цвета Android Auto для ночного вождения. Требуется режим отображения GLES. Действует только при включённом тёмном режиме</string>
+    <string name="aa_desaturation">Обесцвечивание</string>
+    <string name="aa_desaturation_description">Управляет степенью удаления цвета. Действует только при включённом тёмном режиме</string>
+    <string name="gles_required_title">Требуется GLES</string>
+    <string name="gles_required_message">Монохромный режим требует рендеринг GLES для применения фильтра оттенков серого к видеовыходу. Включить GLES?</string>
+    <string name="enable_gles">Включить GLES</string>
     <string name="category_navigation">Навигация</string>
 
     <string name="reconnection_required_title">Требуется переподключение</string>
