+++ app/src/main/res/values-uk/strings.xml
@@ -42,6 +42,14 @@
     <string name="use_gradient_background_description_on_auto">Додаток малює власний фон із центрованим градієнтним колом. Не застосовується до режиму Екстремально темний</string>
     <string name="use_gradient_background_description_off_auto">Використовується класичний фон додатку. Не застосовується до режиму Екстремально темний</string>
     <string name="category_dark_mode">Темний режим</string>
+    <string name="dark_mode_settings">Налаштування темного режиму</string>
+    <string name="aa_monochrome">Монохромний Android Auto</string>
+    <string name="aa_monochrome_description">Зменшує кольори Android Auto для нічного водіння. Потрібен режим відображення GLES. Діє лише при увімкненому темному режимі</string>
+    <string name="aa_desaturation">Знебарвлення</string>
+    <string name="aa_desaturation_description">Визначає, скільки кольору буде видалено. Діє лише при увімкненому темному режимі</string>
+    <string name="gles_required_title">Потрібен GLES</string>
+    <string name="gles_required_message">Монохромний режим потребує рендерингу GLES для застосування фільтра відтінків сірого до відеовиходу. Увімкнути GLES?</string>
+    <string name="enable_gles">Увімкнути GLES</string>
     <string name="category_navigation">Навігація</string>
 
     <string name="reconnection_required_title">Потрібне перепідключення</string>
