+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -41,6 +41,14 @@
     <string name="use_gradient_background_description_on_auto">應用程式繪製自己的背景，漸層圓形置中顯示。不適用於極致深色模式</string>
     <string name="use_gradient_background_description_off_auto">正在使用應用程式的傳統背景。不適用於極致深色模式</string>
     <string name="category_dark_mode">深色模式</string>
+    <string name="dark_mode_settings">深色模式設定</string>
+    <string name="aa_monochrome">將 Android Auto 設為單色</string>
+    <string name="aa_monochrome_description">減少 Android Auto 的色彩以適應夜間駕駛。需要 GLES 檢視模式。僅在深色模式啟用時生效</string>
+    <string name="aa_desaturation">去飽和度</string>
+    <string name="aa_desaturation_description">控制移除多少色彩。僅在深色模式啟用時生效</string>
+    <string name="gles_required_title">需要 GLES</string>
+    <string name="gles_required_message">單色模式需要 GLES 渲染才能在視訊輸出上套用灰階濾鏡。是否啟用 GLES？</string>
+    <string name="enable_gles">啟用 GLES</string>
     <string name="category_navigation">導航</string>
 
     <string name="reconnection_required_title">需要重新連接</string>
