+++ app/src/main/res/values/strings.xml
@@ -46,6 +46,14 @@
     <string name="use_gradient_background_description_on_auto">The app draws its own background with the gradient circle centered. Does not apply to Extreme Dark mode</string>
     <string name="use_gradient_background_description_off_auto">The app\'s classic background is being used. Does not apply to Extreme Dark mode</string>
     <string name="category_dark_mode">Dark Mode</string>
+    <string name="dark_mode_settings">Dark Mode Settings</string>
+    <string name="aa_monochrome">Make Android Auto Monochrome</string>
+    <string name="aa_monochrome_description">Reduces Android Auto colors for night driving. Requires GLES view mode. Only takes effect when dark mode is enabled</string>
+    <string name="aa_desaturation">Desaturation</string>
+    <string name="aa_desaturation_description">Controls how much color is removed. Only takes effect when dark mode is enabled</string>
+    <string name="gles_required_title">GLES Required</string>
+    <string name="gles_required_message">Monochrome mode needs GLES rendering to apply a grayscale filter on the video output. Enable GLES?</string>
+    <string name="enable_gles">Enable GLES</string>
     <string name="category_navigation">Navigation</string>
 
     <string name="reconnection_required_title">Reconnection Required</string>
